requirements.txt
flask
numpy
tensorflow
pillow
flasgger